Refactor paths to use relative ones
