+++ src/components/DiscordMock/DiscordView/style.scss
@@ -1,5 +1,5 @@
-@import "../../../scss/colors";
-@import "../../../scss/theme";
+@import "scss/colors";
+@import "scss/theme";
 
 .discord-view {
     min-height: 100px;
