+++ src/components/DiscordMock/Input/style.scss
@@ -1,6 +1,6 @@
-@import "../../../scss/colors";
-@import "../../../scss/theme";
-@import "../../../scss/util";
+@import "scss/colors";
+@import "scss/theme";
+@import "scss/util";
 
 .discord-input {
     padding: {
