+++ src/components/DiscordMock/Message/style.scss
@@ -1,4 +1,4 @@
-@import "../../../scss/colors";
+@import "scss/colors";
 
 .message {
     line-height: 1.35;
