+++ src/components/DiscordMock/MessageClump/style.scss
@@ -1,4 +1,4 @@
-@import "../../../scss/colors";
+@import "scss/colors";
 
 .message-clump {
     display: flex;
