+++ src/components/DiscordMock/MessageView/style.scss
@@ -1,5 +1,5 @@
-@import "../../../scss/theme";
-@import "../../../scss/util";
+@import "scss/theme";
+@import "scss/util";
 
 %scrollbar {
     &::-webkit-scrollbar {
