+++ src/components/DiscordMock/Reaction/style.scss
@@ -1,4 +1,4 @@
-@import "../../../scss/colors";
+@import "scss/colors";
 
 .mentions-user + .reaction-list {
     .reaction {
