+++ src/components/Footer/style.scss
@@ -1,5 +1,5 @@
-@import "../../scss/colors";
-@import "../../scss/media";
+@import "scss/colors";
+@import "scss/media";
 
 .footer {
     background-color: lighten($-dark, 3%);
