+++ src/components/Header/Links/index.jsx
@@ -2,9 +2,9 @@ import React from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
 import classNames from "classnames";
-import { mapStateToLoggedIn } from "../../../store/reducers/session";
+import { mapStateToLoggedIn } from "store/reducers/session";
 
-import NavLink from "../../NavLink";
+import NavLink from "components/NavLink";
 
 const LoggedInLinks = () => <NavLink to="/home">Home</NavLink>;
