+++ src/components/Header/SessionControl/style.scss
@@ -1,6 +1,6 @@
-@import "../../../scss/colors";
-@import "../../../scss/media";
-@import "../../../scss/util";
+@import "scss/colors";
+@import "scss/media";
+@import "scss/util";
 
 .session-dropdown {
     #session-dropdown-button {
