+++ src/components/Header/style.scss
@@ -1,7 +1,7 @@
-@import '../../scss/colors';
-@import '../../scss/theme';
-@import '../../scss/media';
-@import '../../scss/util';
+@import 'scss/colors';
+@import 'scss/theme';
+@import 'scss/media';
+@import 'scss/util';
 
 .navbar {
     padding: {
