+++ src/components/LoginButton/index.jsx
@@ -1,11 +1,11 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { mapStateToLoggedIn } from "../../store/reducers/session";
+import { mapStateToLoggedIn } from "store/reducers/session";
 
-import Icon from "../Icon";
+import { Link as RouterLink } from "components/Router";
+import Icon from "components/Icon";
 import { Button } from "react-bootstrap";
-import { Link as RouterLink } from "../Router";
 
 import "./style.scss";
