+++ src/components/LoginButton/style.scss
@@ -1,4 +1,4 @@
-@import "../../scss/media";
+@import "scss/media";
 
 .login {
     line-height: 1.2 !important;
