+++ src/components/Placeholder/style.scss
@@ -1,5 +1,5 @@
-@import '../../scss/colors';
-@import '../../scss/animation';
+@import 'scss/colors';
+@import 'scss/animation';
 
 $dark-color: rgba($-dark, 0.1);
 $dark-highlight-color: rgba($-dark, 0.2);
