+++ src/components/UserDisplay/style.scss
@@ -1,4 +1,4 @@
-@import "../../scss/colors";
+@import "scss/colors";
 
 .avatar {
     .avatar-image {
