+++ src/components/Window/style.scss
@@ -1,5 +1,5 @@
-@import "../../scss/colors";
-@import "../../scss/theme";
+@import "scss/colors";
+@import "scss/theme";
 
 %light-chrome {
     .chrome {
