+++ src/pages/Index/style.scss
@@ -1,5 +1,5 @@
-@import "../../scss/media";
-@import "../../scss/theme";
+@import "scss/media";
+@import "scss/theme";
 
 %uppercase-header {
     font-size: 1.2em;
