+++ src/store/api/actions.js
@@ -1,4 +1,4 @@
-import { log } from "../../util";
+import { log } from "utility";
 
 export const API_START = "API_START";
 export const API_END = "API_END";
