+++ src/utility/string.js
